Extracted logic into a util and added unit tests to test it
